backend/fastapi/requirements.txt
fastapi==0.78.0
fastapi-utils==0.2.1
python-dateutil==2.8.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
mysqlclient
pyasn1==0.4.8
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
syncpg==1.1.2
typing_extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
mysql-connector-python==8.0.30
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1